Fix wrong Scheduler used for event processing (#5)
The previously used method as well as Scheduler are just plain wrong

- subscribeOn -> publishOn:
Because subscribeOn does not control the event processing down the pipe

- Schedulers.single() -> Schedulers.parallel():
Events are processed one after another without concurrency issues
on either of them. There is no reason to use single (or elastic),
so parallel should be the correct choice.